tests/wpt/tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-018.html

<!DOCTYPE html>
<html>
<title>Auto-fit repeat tracks with fixed size and auto/explicit item placement</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.masonry {
    display: masonry;
    item-tolerance: 0;
    masonry-direction: row;
    grid-template-rows: repeat(auto-fit, 100px);
    width: 200px;
    height: 1000px;
}

.masonry > div {
    height: 100%;
    width: 100px;
}
</style>
<div class="masonry">
  <div></div>
  <div></div>
  <div style="grid-row: 4"></div>
  <div style="grid-row: 6"></div>
  <div></div>
</div>
<script>
  test(function() {
    const container = document.querySelector('.masonry');
    const computedStyle = window.getComputedStyle(container);
    assert_equals(computedStyle.getPropertyValue('grid-template-rows'),
      "100px 100px 100px 100px 0px 100px 0px 0px 0px 0px");
  });
</script>
</html>
