third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-018-expected.txt

This is a testharness.js-based test.
[FAIL] Auto-fit repeat tracks with fixed size and auto/explicit item placement
  assert_equals: expected "100px 100px 100px 100px 0px 100px 0px 0px 0px 0px" but got "repeat(auto-fit, 100px)"
Harness: the test ran to completion.
